+++ src/pages/home/LoginButtons.vue
@@ -24,6 +24,7 @@ import BaseTextInput from 'components/evm/inputs/BaseTextInput.vue';
 import NativeChainSettings from 'src/antelope/chains/NativeChainSettings';
 import { words } from 'src/pages/home/words';
 import { Subscription } from 'rxjs';
+import { redirectShow, redirect } from 'src/boot/telosCloudJs';
 
 
 export default defineComponent({
@@ -46,81 +47,6 @@ export default defineComponent({
         const globalProps = (getCurrentInstance() as ComponentInternalInstance).appContext.config.globalProperties;
         const isMobile = ref(usePlatformStore().isMobile);
         const isBraveBrowser = ref((navigator as any).brave && (navigator as any).brave.isBrave());
-
-        // Redirect or iFrame -----------------------------------------------------------
-        // If we receive a redirect parameter in the URL, we redirect to the corresponding page after the user has completed the login process
-        // The first thing we do is to check if the URL has a redirect parameter to set a local variable
-        // in case of iframe, we will not redirect the user to the corresponding page,
-        // instead we send the credentials (accountname, email) to the parent window
-        const redirectParam = new URLSearchParams(window.location.search).get('redirect');
-        const iframeParam = new URLSearchParams(window.location.search).get('iframe');
-        const redirect = ref<{url:string, hostname:string} | null>(null);
-        const redirectShow = ref(false);
-        const iframeShow = ref(false);
-        if (redirectParam) {
-            const isValid = new RegExp('^(http|https)://', 'i').test(redirectParam);
-            if (isValid) {
-                redirect.value = {
-                    url: redirectParam,
-                    hostname: new URL(redirectParam).hostname,
-                };
-                redirectShow.value = true;
-            }
-        }
-
-        if (iframeParam) {
-            const isValid = new RegExp('^(http|https)://', 'i').test(iframeParam);
-            if (isValid) {
-                redirect.value = {
-                    url: iframeParam,
-                    hostname: new URL(iframeParam).hostname,
-                };
-                iframeShow.value = true;
-            }
-        }
-
-        const subscription = ant.events.onLoggedIn.subscribe({
-            next: () => {
-                subscription.unsubscribe();
-                // if the redirect parameter is present, show a confirm notification to the user
-
-                if (redirectShow.value && redirect.value) {
-                    const hostname = redirect.value.hostname;
-                    const message = globalProps.$t('home.redirect_notification_message', { hostname });
-                    globalProps.$notifyWarningWithAction(message, {
-                        label: ant.config.localizationHandler('home.redirect_me'),
-                        handler: () => {
-                            // we redirect the user to the url
-                            if (redirect.value) {
-                                // we need to generate a new url based on redirect.value?.url adding the accountStore.loggedNativeAccount?.account
-                                // and the email of the user if it's a metakeep authenticator
-                                const url = new URL(redirect.value.url);
-                                url.searchParams.set('account', accountStore.loggedNativeAccount?.account || '');
-                                const authenticator = accountStore.loggedNativeAccount.authenticator;
-                                console.log('adding the email...'); // FIXME: remove this line
-                                const auth = authenticator as never as MetakeepAuthenticator;
-                                url.searchParams.set('email', auth.getEmail());
-                                console.log('redirecting to', url.toString()); // FIXME: remove this line
-                                window.location.href = url.toString();
-                            }
-                        },
-                    });
-                } else if (iframeShow.value) {
-                    const authenticator = accountStore.loggedNativeAccount.authenticator;
-                    const auth = authenticator as never as MetakeepAuthenticator;
-                    const credentials: {account: string, email: string, keys: string[] } = {
-                        account: accountStore.loggedNativeAccount?.account || '',
-                        email: auth.getEmail(),
-                        keys: auth.getKeys() ?? [],
-                    };
-                    const str = JSON.stringify(credentials);
-                    console.log('credentials', credentials, str); // FIXME: remove this line
-                    accountStore.logout();
-                    window.parent.postMessage(str, '*');
-                }
-            },
-        });
-
 
         const showEVMButtons = computed(() =>
             props.chain === 'evm');
